requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases
pyodbc
alembic
sqlalchemy[asyncio]
asyncpg 
aioodbc 
pydantic-settings